refactor: simplify

+++ src/code/conv/conv_3ntt.hpp
@@ -1,7 +1,7 @@
 #ifndef TIFALIBS_CONV_CONV_3NTT
 #define TIFALIBS_CONV_CONV_3NTT
 
-#include "../math/mul_mod_u.hpp"
+#include "../math/mul_mod.hpp"
 #include "conv_dft.hpp"
 #include "ntt.hpp"
 
@@ -10,8 +10,8 @@ namespace tifa_libs::math {
 // 167772161, 469762049, 754974721
 template <class mint0, class mint1, class mint2>
 CEXP vecu64 conv_3ntt_u64(std::tuple<NTT<mint0>, NTT<mint1>, NTT<mint2>> &ntt3, vecu64 CR l, vecu64 CR r, u64 mod, u32 ans_size = 0) {
-  CEXP u32 m0 = mint0::mod(), m1 = mint1::mod(), m2 = mint2::mod();
-  const u32 r01 = mint1(m0).inv().val(), r02 = mint2(m0).inv().val(), r12 = mint2(m1).inv().val(), r02r12 = (u32)mul_mod_u(r02, r12, m2);
+  CEXP u64 m0 = mint0::mod(), m1 = mint1::mod(), m2 = mint2::mod();
+  const u64 r01 = mint1(m0).inv().val(), r02 = mint2(m0).inv().val(), r12 = mint2(m1).inv().val(), r02r12 = (u32)mul_mod_u(r02, r12, m2);
   const u64 w1 = m0 % mod, w2 = mul_mod_u(m0, m1, mod);
 
   if (!ans_size) ans_size = u32(l.size() + r.size() - 1);
